Fix memory recorder (#6952)

+++ source/source_base/memory.h
@@ -45,7 +45,7 @@ class Memory
      */
     static void record(
       const std::string &name_in,
-      const size_t &n_in,
+      const long &n_in,
       const bool accumulate = false
     );
